Define Move sealed interface and classes
Create a Move interface to define shogi moves. Moves can be applied to a
Position with the extension function `Position.doMove(move)` to return the
Position after the move, without any constraint on the legality or validity of
the move.

Move depends on KomaPromotions to be able to do promotion moves. KomaPromotions
will be defined separate from the rest of the shogi rules.

Unit tests for `doMove()` are included.