Refactor DestinationGeneration into KomaMovement and its definitions
DestinationGeneration was used to generate valid destination squares, given a
board state and a particular koma.

The new class `KomaMovement` describes the generalised movement of a koma, and
the method `.getDestinations(...)` takes the board state and calculates the
valid destination squares based on the koma movement defined.

The definitions for the standard shogi koma are in `KomaMovementDefinitions`.